move non max suppresion tests to own files
