+++ tfjs-core/src/tests.ts
@@ -80,7 +80,6 @@ import './ops/gather_nd_test';
 import './ops/gram_schmidt_test';
 import './ops/greater_equal_test';
 import './ops/greater_test';
-import './ops/image_ops_test';
 import './ops/in_top_k_test';
 import './ops/leaky_relu_test';
 import './ops/less_equal_test';
@@ -98,6 +97,8 @@ import './ops/max_pool_test';
 import './ops/max_pool_with_argmax_test';
 import './ops/moving_average_test';
 import './ops/multinomial_test';
+import './ops/non_max_suppression_async_test';
+import './ops/non_max_suppression_test';
 import './ops/not_equal_test';
 import './ops/one_hot_test';
 import './ops/operation_test';
